🚚 move Habit type to App.tsx
